refactor: moved entities from commons to server
