+++ server/src/main/java/com/meowny/server/entity/BaseAuditEntity.java
@@ -1,4 +1,4 @@
-package com.meowny.commons.entity;
+package com.meowny.server.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.MappedSuperclass;
