+++ server/src/main/java/com/meowny/server/entity/Transaction.java
@@ -1,21 +1,21 @@
-package com.meowny.commons.entity;
+package com.meowny.server.entity;
 
 import jakarta.persistence.*;
 import jakarta.validation.constraints.*;
 import java.math.BigDecimal;
 import java.time.LocalDate;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Objects;
 
 @Entity
-@Table(name = "recurring_transactions",
+@Table(
+        name = "transactions",
         indexes = {
-                @Index(name = "idx_recurring_tx_worker", columnList = "is_active, next_due_date"),
-                @Index(name = "idx_recurring_tx_user", columnList = "user_id")
+                @Index(name = "idx_tx_user_type_date", columnList = "user_id, type, payment_date"),
+                @Index(name = "idx_tx_category", columnList = "category_id"),
+                @Index(name = "idx_tx_template", columnList = "recurring_transaction_id")
         }
 )
-public class RecurringTransaction extends BaseAuditEntity {
+public class Transaction extends BaseAuditEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -23,12 +23,12 @@ public class RecurringTransaction extends BaseAuditEntity {
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false)
-    @NotNull(message = "Template must be assigned to a user")
+    @NotNull(message = "Transaction must be assigned to a user")
     private User user;
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false, length = 10)
-    @NotNull(message = "Type is required")
+    @NotNull(message = "Transaction type is required")
     private TransactionType type;
 
     @ManyToOne(fetch = FetchType.LAZY)
@@ -36,65 +36,64 @@ public class RecurringTransaction extends BaseAuditEntity {
     @NotNull(message = "Category is required")
     private Category category;
 
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "recurring_transaction_id")
+    private RecurringTransaction sourceTemplate;
+
     @Column(length = 50, nullable = false)
     @NotBlank(message = "Name is required")
     @Size(max = 50, message = "Name must be 50 characters or fewer")
     private String name;
 
-    @Column(nullable = false, precision = 12, scale = 2)
     @NotNull(message = "Amount is required")
+    @Column(nullable = false, precision = 12, scale = 2)
     @PositiveOrZero(message = "Amount must be zero or positive")
     @Digits(integer = 12, fraction = 2)
     private BigDecimal amount;
 
-    @Column(name = "next_due_date", nullable = false)
-    @NotNull(message = "Next due date is required")
-    private LocalDate nextDueDate;
+    @Column(name = "payment_date", nullable = false)
+    @NotNull(message = "Payment date is required")
+    @PastOrPresent(message = "Payment date cannot be in the future")
+    private LocalDate paymentDate;
 
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    @NotNull(message = "Frequency is required")
-    private Frequency frequency;
+    @Column(length = 100)
+    @Size(max = 100, message = "Description must be 100 characters or fewer")
+    private String description;
 
-    @Column(name = "is_active", nullable = false)
-    private boolean isActive = true;
+    public Transaction() {}
 
-    @OneToMany(mappedBy = "sourceTemplate", cascade = CascadeType.ALL)
-    private List<Transaction> history = new ArrayList<>();
-
-    public RecurringTransaction() {}
-
-    // Getters and Setters...
     public Long getId() { return id; }
     public void setId(Long id) { this.id = id; }
+
     public User getUser() { return user; }
     public void setUser(User user) { this.user = user; }
+
     public TransactionType getType() { return type; }
     public void setType(TransactionType type) { this.type = type; }
+
     public Category getCategory() { return category; }
     public void setCategory(Category category) { this.category = category; }
+
+    public RecurringTransaction getSourceTemplate() { return sourceTemplate; }
+    public void setSourceTemplate(RecurringTransaction sourceTemplate) { this.sourceTemplate = sourceTemplate; }
+
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
+
     public BigDecimal getAmount() { return amount; }
     public void setAmount(BigDecimal amount) { this.amount = amount; }
-    public LocalDate getNextDueDate() { return nextDueDate; }
-    public void setNextDueDate(LocalDate nextDueDate) { this.nextDueDate = nextDueDate; }
-    public Frequency getFrequency() { return frequency; }
-    public void setFrequency(Frequency frequency) { this.frequency = frequency; }
-    public boolean isActive() { return isActive; }
-    public void setActive(boolean active) { isActive = active; }
-    public List<Transaction> getHistory() { return history; }
 
-    public void addTransaction(Transaction tx) {
-        this.history.add(tx);
-        tx.setSourceTemplate(this);
-    }
+    public LocalDate getPaymentDate() { return paymentDate; }
+    public void setPaymentDate(LocalDate paymentDate) { this.paymentDate = paymentDate; }
+
+    public String getDescription() { return description; }
+    public void setDescription(String description) { this.description = description; }
 
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-        RecurringTransaction that = (RecurringTransaction) o;
+        Transaction that = (Transaction) o;
         return id != null && Objects.equals(id, that.id);
     }
